Avoid ARQ when resolving sources and jvadoc for IDEs
ArtifactResolutionQuery is old, discouraged, and serialized. We should migrate to a modern
implementation based on artifactView.

This PR attempts to make that change. However, there are existing issues with artifactView
variant reselection that prevent dependencies with classifiers from getting reselected.

A reproducer for this bug is included in this commit. That bug must be resolved before
this effort can continue